+++ src/main/java/pl/farmapp/backend/repository/FarmerDetailsRepository.java
@@ -5,9 +5,9 @@ import pl.farmapp.backend.entity.FarmerDetails;
 
 import java.util.Optional;
 
-public interface FarmerDetailsRepository extends JpaRepository<FarmerDetails, Long> {
+public interface FarmerDetailsRepository extends JpaRepository<FarmerDetails, Integer> {
 
-    Optional<FarmerDetails> findByFarmerId(Long farmerId);
+    Optional<FarmerDetails> findByFarmerId(Integer farmerId);
 
-    boolean existsByFarmerId(Long farmerId);
+    boolean existsByFarmerId(Integer farmerId);
 }
